Try to compile the files before compiling them all
There are some yara files which the yara-python library won't accept as
valid syntax. There's no documentation as to which ones these will be or
why so I've gone for this brute force approach.

It attempts to compile each file individually, if it works, it allows it
through to the list which are all compiled together. If it fails, it's
removed.

It's very inefficient doing everything twice but it works.