Send the file id
Because the s3 key will be in the format cognitoId/fileId, this splits
the key on the / and takes the last item. This works no matter how many
subdirectories or for none so it should be reasonably future proof so
long as the file is always named with the file id.

Also added a test for it.