Change to delete successful messages
If we get more than one message into the lambda, and one fails and one
succeeds, the successful message will be deleted off the queue to
prevent reprocessing and then the lambda fails to allow the failed
messages to requeue.
There are tests using a mock s3 library as well.