Refactor to put as much code in the common module as possible.
For new lambdas, it should be possible to create a new instance of the
Processor class with the correct type parameters and call process.

This may be a way in the future to have a single lambda which does all
of the api updates but for now, this will make writing the separate ones
easier.